+++ web/src/app/data-templates/outputs.ts
@@ -24,6 +24,7 @@ import * as postgresTemplate from './outputs/postgres.json';
 import * as printTemplate from './outputs/print.json';
 import * as redisTemplate from './outputs/redis.json';
 import * as textTemplate from './outputs/text.json';
+import * as xmlTemplate from './outputs/xml.json';
 
 
 export const outputs = [
@@ -44,7 +45,8 @@ export const outputs = [
     postgresTemplate,
     redisTemplate,
     printTemplate,
-    textTemplate
+    textTemplate,
+    xmlTemplate
 ];
 
 /*********************** */